Distribute components into separate dirs
