+++ src/components/controls/IncrementDecrement/__tests__/smoke.js
