+++ src/components/controls/MonthSelector/__tests__/smoke.js
