+++ src/components/controls/MonthSelector/package.json
