+++ src/components/controls/ScrollingSelector/ScrollingSelector.css
