+++ src/components/controls/ScrollingSelector/__tests__/smoke.js
