+++ src/components/controls/ThrottledInputRange/ThrottledInputRange.js
