+++ src/components/controls/ThrottledInputRange/__tests__/smoke.js
