+++ src/components/controls/VariableSelector/__tests__/smoke.js
