+++ src/components/controls/YearSelector/__tests__/smoke.js
