+++ src/components/main/App/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Container } from 'react-bootstrap';
 
-import logger from '../../logger';
+import logger from '../../../logger';
 import Header from '../Header';
 import Tool from '../Tool'
