+++ src/components/main/App/__tests__/smoke.js
