+++ src/components/main/App/package.json
