+++ src/components/main/Header/__tests__/smoke.js
