+++ src/components/main/Tool/__tests__/smoke.js
