+++ src/components/map/ColourScale/ColourScale.js
