+++ src/components/map/DataMap/__tests__/smoke.js
