+++ src/components/map/StationPopup/__tests__/smoke.js
