+++ src/components/misc/VariableLabel/VariableLabel.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { variableLabels } from '../../utils/variables';
+import { variableLabels } from '../../../utils/variables';
 
 export default function VariableLabel({ variable }) {
   return variableLabels[variable] || `${variable}`;
